requirements/requirements-dev.txt
-r requirements.txt
django-stubs==5.0.2
djangorestframework-stubs==3.15.0
mypy==1.10.0
pre-commit==3.7.1
types-PyMySQL==1.1.0.20240524
